Initial version of "scrabble" game.
This "game" is not really a game.  It is just a test for the Library files WordBoard.h/cpp an WordValidator.h/cpp and the support text file WordList.txt that holds a huge list of valid words.  Note that all words must be in upper case with no spaces or punctuation ('-' dash marks are allowed but must match exactly).